Use user's Library/Developer/org.jenkins-ci.slave.jnlp directory instead of user's home for scripts and Jenkins workspace
Also slave.jnlp.sh tests if it was launched as agent or daemon and
calls security.sh with(out) `--agent` as appropriate. If security.sh is
called with `--agent`, it will not (un)lock the login.keychain.
Note that install.sh does not, as of yet, support installing slave to
run as agent.